Fixed issue if you want to do more than one request with the same request type. include leads to an error like this: Fatal error: Cannot redeclare class OAuth_Request_Resource in /var/www/../OAuth_Request_Resource.php on line 18
After doing the get_user_info request via the flickr provider, I needed to make another request to get a list of photosets and for that purpose I added another method to the flickr provider which does another request with type 'resource' and that's when I got the above error because the OAuth_Request forge tries to inlcude the OAuth_Request_Resource.php again and that leads to the redeclaration of the class.

There might be other cases(classes where this is an issue. I could update all classes to use include_once instead of include, don't think it would break anything.